refactor: keep e2e tests unchanged
