+++ tests/e2e/config/mod.rs
@@ -1,3 +1,1 @@
-mod edit;
-mod prompt_format;
-mod xdg;
+mod config;
