+++ tests/e2e/daemon/mod.rs
@@ -1,4 +1,2 @@
-mod concurrency;
 mod lifecycle;
-mod restart;
 mod watch;
