+++ tests/e2e/helpers/mod.rs
@@ -1,7 +1,6 @@
 mod daemon_handle;
 mod env;
 mod multi_repo;
-mod repo;
 
 pub use daemon_handle::{DaemonHandle, FakeDaemonHandle};
 pub use env::TestEnv;
